Create README.md
minecraft should be opened in a small window on the right and centered, its edge should be touching the endge of the screen for this to work, 
since i used coordinates of mouse pointer on screen to determine where to click. Ik its not a very good method, but its my first time creating a bot like this. I'm still a student learning and experimenting
with python.